WIP: Improve CodeDeploy Appspec
- Add database migration script invocation to Appspec
- Add npm build script invocation to Appspec
- Add scripts for the above
- Add if-clauses to prevent errors if TM systemd unit is absent.
- Add tags to instances at ASG

+++ scripts/aws/cloudformation/service-stop.sh
@@ -1,4 +1,11 @@
 #!/usr/bin/env bash
 
-/bin/systemctl stop tasking-manager.service
-/bin/sleep 10
+printf "Stopping tasking-manager service..."
+
+if [[ ! -f /etc/systemd/system/tasking-manager.service ]]; then
+  /bin/systemctl stop tasking-manager.service
+  /bin/sleep 10
+  printf "Tasking-manager service stopped"
+else
+  printf "Tasking-manager service does not exist; Failed."
+fi
